requirements.txt
python-decouple>=3.3
watchdog>=2.0.2